+++ src/main/java/org/mathevaluator/util/ExpressionCleaner.java
@@ -5,20 +5,28 @@ import static org.springframework.util.StringUtils.countOccurrencesOf;
 public class ExpressionCleaner {
 
     public static String cleanSpaces(String expression) {
-	return expression.replaceAll("\\s", "");
+        return expression.replaceAll("\\s", "");
     }
 
     public static String cleanParentheses(String expression) {
-	String f = expression;
-	if ((expression.length() > 2) && f.startsWith("(") && f.endsWith(")")
-		&& countOccurrencesOf(expression, "(") == countOccurrencesOf(expression, ")")) {
-	    f = f.substring(1, f.length() - 1);
-	}
-	if (!f.equals(expression)) {
-	    return cleanParentheses(f);
-	} else {
-	    return f;
-	}
+        return cleanDelimiter(expression, "(", ")");
+    }
+
+    public static String cleanBrackets(String expression) {
+        return cleanDelimiter(expression, "[", "]");
+    }
+
+    public static String cleanDelimiter(String expression, String delimiterBegin, String delimiterEnd) {
+        String f = expression;
+        if ((expression.length() > 2) && f.startsWith(delimiterBegin) && f.endsWith(delimiterEnd)
+                && countOccurrencesOf(expression, delimiterBegin) == countOccurrencesOf(expression, delimiterEnd)) {
+            f = f.substring(1, f.length() - 1);
+        }
+        if (!f.equals(expression)) {
+            return cleanParentheses(f);
+        } else {
+            return f;
+        }
     }
 
 }